Restructure E2E tests
